♻️: refactor RedisTemplate wrapping 적용, class 명 수정 #14
