+++ plu-api/src/main/kotlin/com/th/plu/api/config/interceptor/LoginCheckHandler.kt
@@ -1,21 +1,21 @@
 package com.th.plu.api.config.interceptor
 
+import com.th.plu.api.service.auth.jwt.JwtHandler
 import com.th.plu.common.exception.code.ErrorCode
 import com.th.plu.common.exception.model.UnauthorizedException
-import com.th.plu.common.util.JwtUtils
 import jakarta.servlet.http.HttpServletRequest
 import org.springframework.stereotype.Component
 
 @Component
 class LoginCheckHandler(
-    private val jwtUtils: JwtUtils
+    private val jwtHandler: JwtHandler
 ) {
     fun getMemberId(request: HttpServletRequest): Long {
         val bearerToken: String? = request.getHeader("Authorization")
         if (!bearerToken.isNullOrBlank() && bearerToken.startsWith("Bearer ")) {
             val accessToken = bearerToken.substring("Bearer ".length)
-            if (jwtUtils.validateToken(accessToken)) {
-                val memberId = jwtUtils.getMemberIdFromJwt(accessToken)
+            if (jwtHandler.validateToken(accessToken)) {
+                val memberId = jwtHandler.getMemberIdFromJwt(accessToken)
                 if (memberId != null) {
                     return memberId
                 }
