+++ plu-api/src/main/kotlin/com/th/plu/api/service/auth/CommonAuthService.kt
@@ -1,7 +1,7 @@
 package com.th.plu.api.service.auth
 
+import com.th.plu.api.service.auth.jwt.JwtHandler
 import com.th.plu.api.service.member.MemberServiceUtils
-import com.th.plu.common.util.JwtUtils
 import com.th.plu.domain.domain.member.repository.MemberRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -9,13 +9,13 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class CommonAuthService(
     private val memberRepository: MemberRepository,
-    private val jwtUtils: JwtUtils
+    private val jwtHandler: JwtHandler
 ) {
 
     @Transactional
     fun logout(memberId: Long) {
         val member = MemberServiceUtils.findMemberById(memberRepository, memberId)
-        jwtUtils.expireRefreshToken(member.id!!)
+        jwtHandler.expireRefreshToken(member.id!!)
         member.resetFcmToken()
     }
 }
