+++ plu-common/build.gradle.kts
@@ -15,14 +15,6 @@ dependencies {
 
     // web
     implementation("org.springframework.boot:spring-boot-starter-web")
-
-    // Redis
-    implementation("org.springframework.boot:spring-boot-starter-data-redis")
-
-    //jwt
-    implementation("io.jsonwebtoken:jjwt-api:0.11.2")
-    runtimeOnly("io.jsonwebtoken:jjwt-impl:0.11.2")
-    runtimeOnly("io.jsonwebtoken:jjwt-jackson:0.11.2")
 }
 repositories {
     mavenCentral()
